requirements.txt
azure-functions
psycopg2-binary 
requests
